+++ php-laraval-app/app/Models/Product.php
@@ -3,7 +3,6 @@
 namespace App\Models;
 
 use App\Enums\Discount\CustomerDiscountTypeEnum;
-use App\Enums\MembershipTypeEnum;
 use App\Http\Traits\TranslationsTrait;
 use Auth;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -163,9 +162,9 @@ class Product extends Model
         return $this->morphToMany(Subscription::class, 'itemable', 'subscription_items');
     }
 
-    public function scheduleCleanings(): MorphToMany
+    public function schedules(): MorphToMany
     {
-        return $this->morphToMany(ScheduleCleaning::class, 'itemable', 'schedule_cleaning_items');
+        return $this->morphToMany(Schedule::class, 'itemable', 'schedule_items');
     }
 
     public function stores(): BelongsToMany